Some refactoring of categoryvalues
