+++ src/EA.Weee.Core/AatfReturn/NonObligatedCategoryValue.cs
@@ -4,18 +4,11 @@
     using Helpers;
     using Validation;
 
-    public class NonObligatedCategoryValue
+    public class NonObligatedCategoryValue : CategoryValue
     {
-        public string CategoryDisplay { get; set; }
-
-        public int CategoryId { get; set; }
 
         [TonnageValue("CategoryId")]
         public string Tonnage { get; set; }
-
-        public string HouseHold { get; set; }
-
-        public string NonHouseHold { get; set; }
 
         public bool Dcf { get; set; }
 
@@ -23,10 +16,8 @@
         {
         }
 
-        public NonObligatedCategoryValue(WeeeCategory category)
+        public NonObligatedCategoryValue(WeeeCategory category) : base(category)
         {
-            CategoryDisplay = category.ToDisplayString();
-            CategoryId = (int)category;
         }
     }
 }
