+++ src/EA.Weee.Core/AatfReturn/ObligatedCategoryValue.cs
@@ -1,0 +1,17 @@
+﻿namespace EA.Weee.Core.AatfReturn
+{
+    using DataReturns;
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Text;
+    using System.Threading.Tasks;
+    
+    public class ObligatedCategoryValue : CategoryValue
+    {
+        public ObligatedCategoryValue(WeeeCategory category) : base(category)
+        {
+
+        }
+    }
+}
